feat: complete game systems — starmap, combat, colonies, admin, holonet
Major feature additions:
- Starmap: galaxy generation, system exploration, planet surfaces,
  wormholes, influence areas, faction zones
- Spacecraft: cargo system, fleet management, transfer service
- Colonies: expanded building/resource system, colony seeding
- HoloNet: comments, ratings, checkpoints
- Admin panel: starmap editor, ship class management
- Onboarding: claim-colony flow with sector/system/planet selection
- Settings module, research unlock resolver
- Frontend: admin pages, starmap-admin store, spacecraft components

+++ apps/backend/src/modules/tick/entities/game-tick-state.entity.ts
@@ -23,7 +23,7 @@ export class GameTickState {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column({ type: 'int' })
+  @Column({ type: 'bigint' })
   tickNumber: number;
 
   @Column({ type: 'varchar' })
